requirements.txt
openai
anthropic
google-genai
python-dotenv
pydantic
tqdm